+++ html_output/hy_world_2_0/TUTORIAL_CHANGELOG.md
@@ -4,6 +4,47 @@
 
 每个版本必须包含“模块变更明细”：逐项写出模块编号、模块名称、组件文件、新增或调整的具体内容，以及需要审查的状态变化。若没有修改运行模块，也必须明确写“无模块变更”，不能只写笼统的功能概述。
 
+## V0.58「启程星册·首章导览姬」
+
+日期：2026-08-17
+
+### 修改目标
+
+- 全文速览不再占用封面空间，封面只承担论文定位、新旧方法对照和开始学习入口。
+- 把全文主线图放到第一章最前端，使读者正式进入教程后先获得系统全貌，再学习世界模型定义与四阶段管线。
+- 保留 Quick Read 的完整八站专家导航能力，同时让入口与速览内容处于同一个教学位置。
+
+### 模块变更明细
+
+- 新增模块 1.0“全文速览：为什么做、怎么做、做到什么”，组件 `src/modules/hy-paper-overview.tsx`：承接 V0.57 的三条论文主张、五组件链和结果边界栏。
+- 模块 1.0 在普通教程中新增“进入全文专家导航”入口；进入 Quick Read 后隐藏该入口，避免重复跳转。
+- 调整 Hero 论文专属组件 `src/modules/hy-hero.tsx`：删除普通模式下的速览入口，删除 Quick Read 模式下嵌入封面的速览正文；Hero 仅保留固定专家导航的挂载职责。
+- 调整 `src/data/tutorial.ts`：将模块 1.0 插入第一章模块列表首位，明确位于 1.1“什么样的输出才算一个可用世界？”之前。
+- 调整 `src/modules/registry.tsx`：注册独立组件 `hy-paper-overview`，使速览从 Hero 组件职责中拆分出来。
+- 调整 `src/styles/paper.css`：速览进入标准模块容器后移除多余外边距和顶部边线；专家导航入口与模块宽度对齐。
+- 更新模块 8.2 `src/modules/hy-update-log.tsx`：最近三版改为 V0.58、V0.57、V0.56，默认选中 V0.58。
+
+### 事实与教学边界
+
+- 本次只改变速览的页面位置和组件职责，不修改五组件链、5.60 秒、712 秒或高斯数量等论文事实。
+- `quick-overview` 仍是专家导航第一落点，但实际 DOM 位置已从封面移到第一章模块 1.0。
+
+### 验证结果
+
+- 教程生产构建通过：70 个模块完成转换；仅保留既有的单包体积提示，不影响构建产物。
+- `paper-skill` 结构验证通过：8 章、20 个主动模块、22 个组件全部注册，无占位符残留。
+- 普通模式封面检查通过：Hero 内不存在速览正文、速览入口或固定专家导航，页面无横向溢出。
+- 第一章顺序检查通过：模块标题依次为 1.0、1.1、1.2、1.3，速览组件位于第一章第一个模块内部。
+- Quick Read 首落点检查通过：入口跳转到 `?quickread=1#quick-overview`，目标位于第一章且顶部保留约 18px 安全区；Hero 中没有速览副本。
+- 窄屏检查通过：固定导航默认收起，模块 1.0 无横向溢出；Quick Read 模式中不重复显示“进入全文专家导航”入口。
+- 页面重新加载后未发现新增运行时错误。
+
+### 审查重点
+
+- 普通模式下封面不应出现“快速展示”“全文概述”或五组件速览内容。
+- 开始学习后，第一章应先显示模块 1.0 全文速览，再依次显示 1.1、1.2 和 1.3。
+- 从模块 1.0 进入 Quick Read 后，第一落点仍应准确回到模块 1.0，而不是封面。
+
 ## V0.57「一屏星图·极速讲解姬」
 
 日期：2026-08-16
