+++ html_output/hy_world_2_0/src/modules/hy-hero.tsx
@@ -18,29 +18,6 @@ const quickReadStops = [
   { id: 'quick-conclusion', chapter: '结论边界', title: '可运行世界，但仍是离线生产', cue: '最终产物可保存、渲染和交互；WorldMirror 子步骤可到 5.60 秒，但完整世界生成仍约 712 秒。', components: '128 views: 5.60 s · End-to-end: 712 s · NVIDIA H20' },
 ] as const;
 
-const overviewStages = [
-  ['01', 'HY-Pano 2.0', '把文本或单图扩展为 360° 世界种子'],
-  ['02', 'WorldNav', '解析场景并规划五类补盲轨迹'],
-  ['03', 'WorldStereo 2.0', '用关键帧、双记忆和四步采样扩展观察'],
-  ['04', 'WorldMirror 2.0', '一次共享前向恢复五类三维产物'],
-  ['05', '3DGS + WorldLens', '压缩显式资产并接入光照、碰撞与漫游'],
-] as const;
-
-function PaperOverview() {
-  return <section id="quick-overview" className="hy-paper-overview" aria-label="HY-World 2.0 全文快速概述">
-    <header><span>全文最快概述</span><strong>生成补观察，重建守几何，最终交付可运行三维世界</strong></header>
-    <div className="hy-overview-theses">
-      <article><b>为什么做</b><p>视频生成擅长想象未见区域，却不稳定保存三维状态；多视图重建忠实恢复已见几何，却无法补齐盲区。</p></article>
-      <article><b>核心办法</b><p>文本与单图先走生成链，多视图与视频可直接重建；两条路线最终都由 WorldMirror 2.0 凝结为显式资产。</p></article>
-      <article><b>做成什么</b><p>输出可保存、可换视角渲染、可压缩，并能进入 WorldLens 支持重新照明、碰撞和角色漫游。</p></article>
-    </div>
-    <ol className="hy-overview-pipeline">
-      {overviewStages.map(([order, name, role]) => <li key={name}><span>{order}</span><div><strong>{name}</strong><small>{role}</small></div></li>)}
-    </ol>
-    <footer><b>结果与边界</b><span>3DGS 从 6.000M 压到 1.381M；WorldMirror 在 H20 四卡、128 视图条件下为 5.60 秒；完整世界生成约 712 秒，仍是离线流程。</span></footer>
-  </section>;
-}
-
 const waitForQuickReadFrame = (delay: number) => new Promise<void>((resolve) => window.setTimeout(resolve, delay));
 
 async function revealQuickReadTarget(targetId: string) {
@@ -87,9 +64,7 @@ function QuickReadNavigator() {
     };
   }, [quickMode]);
 
-  if (!quickMode) {
-    return <a className="quick-read-entry" href="?quickread=1#quick-overview"><span>快速展示</span><strong>全文总览与专家导航</strong><b>→</b></a>;
-  }
+  if (!quickMode) return null;
 
   const goTo = (targetId: string) => {
     setActiveId(targetId);
@@ -97,7 +72,7 @@ function QuickReadNavigator() {
     if (window.matchMedia('(max-width: 620px)').matches) setCollapsed(true);
   };
 
-  return <><PaperOverview /><aside className={`quick-read-panel ${collapsed ? 'collapsed' : ''}`} aria-label="Quick Read 导航">
+  return <aside className={`quick-read-panel ${collapsed ? 'collapsed' : ''}`} aria-label="Quick Read 导航">
       <header>
         <div><span>QUICK READ</span><strong>专家展示导航</strong><small>完整论文八个讲解落点</small><a href="./">返回完整教程</a></div>
         <button type="button" onClick={() => setCollapsed((value) => !value)} aria-label={collapsed ? '展开 Quick Read 导航' : '收起 Quick Read 导航'} title={collapsed ? '展开导航' : '收起导航'}>{collapsed ? '＋' : '−'}</button>
@@ -106,7 +81,7 @@ function QuickReadNavigator() {
         <nav>{quickReadStops.map((item) => <button key={item.id} type="button" className={active.id === item.id ? 'selected' : ''} aria-current={active.id === item.id ? 'step' : undefined} onClick={() => goTo(item.id)}><span>{item.chapter}</span><strong>{item.title}</strong></button>)}</nav>
         <section aria-live="polite"><span>当前讲解提示</span><strong>{active.title}</strong><p>{active.cue}</p><small title={active.components}>{active.components}</small></section>
       </>}
-    </aside></>;
+    </aside>;
 }
 
 function CanvasView({ width = 560, height = 240, animate = false, draw }: { width?: number; height?: number; animate?: boolean; draw: (ctx: CanvasRenderingContext2D, time: number) => void }) {
