+++ html_output/hy_world_2_0/src/modules/registry.tsx
@@ -15,6 +15,7 @@ import { HyMissionPlanner } from './hy-mission-planner';
 import { HyModelEvolution } from './hy-model-evolution';
 import { HyPanorama } from './hy-panorama';
 import { HyPanoramaCuration } from './hy-panorama-curation';
+import { HyPaperOverview } from './hy-paper-overview';
 import { HyPerformanceCompare } from './hy-performance-compare';
 import { HyResolution } from './hy-resolution';
 import { HyTrainingStages } from './hy-training-stages';
@@ -45,6 +46,7 @@ widgetRegistry['hy-mission-planner'] = HyMissionPlanner;
 widgetRegistry['hy-model-evolution'] = HyModelEvolution;
 widgetRegistry['hy-panorama'] = HyPanorama;
 widgetRegistry['hy-panorama-curation'] = HyPanoramaCuration;
+widgetRegistry['hy-paper-overview'] = HyPaperOverview;
 widgetRegistry['hy-performance-compare'] = HyPerformanceCompare;
 widgetRegistry['hy-resolution'] = HyResolution;
 widgetRegistry['hy-training-stages'] = HyTrainingStages;
