+++ html_output/hy_world_2_0/src/styles/paper.css
@@ -8801,6 +8801,8 @@
 .hy-paper-overview > footer { display: grid; gap: 3px; padding: 8px 9px; border-left: 3px solid var(--orange); background: #fff8e8; }
 .hy-paper-overview > footer b { color: var(--orange); font-size: 11px; }
 .hy-paper-overview > footer span { color: var(--ink-2); font-size: 11px; line-height: 1.45; }
+.module-body .hy-paper-overview { margin: 0; padding-top: 0; border-top: 0; }
+.hy-overview-entry { margin: 2px 0 0; }
 
 .quick-read-panel { position: fixed; z-index: 120; left: 14px; right: 14px; top: 10px; display: grid; grid-template-columns: 190px minmax(0,1fr) 260px; width: auto; min-height: 92px; overflow: hidden; border: 1px solid #9ba8b8; background: rgba(255,255,255,.98); box-shadow: 0 10px 28px rgba(33,50,74,.17); }
 .quick-read-panel > header { display: flex; align-items: center; justify-content: space-between; gap: 10px; padding: 10px 11px; border-right: 1px solid var(--line); background: #fff; }
